+++ cmd/dbaas_update.go
@@ -25,6 +25,7 @@ type dbaasServiceUpdateCmd struct {
 	HelpMysql             bool   `cli-usage:"show usage for flags specific to the mysql type"`
 	HelpPg                bool   `cli-usage:"show usage for flags specific to the pg type"`
 	HelpRedis             bool   `cli-usage:"show usage for flags specific to the redis type"`
+	HelpValkey            bool   `cli-usage:"show usage for flags specific to the valkey type"`
 	MaintenanceDOW        string `cli-flag:"maintenance-dow" cli-usage:"automated Database Service maintenance day-of-week"`
 	MaintenanceTime       string `cli-usage:"automated Database Service maintenance time (format HH:MM:SS)"`
 	Plan                  string `cli-usage:"Database Service plan"`
@@ -99,6 +100,18 @@ type dbaasServiceUpdateCmd struct {
 	RedisMigrationDBName    string   `cli-flag:"redis-migration-dbname" cli-usage:"database name for bootstrapping the initial connection" cli-hidden:""`
 	RedisMigrationMethod    string   `cli-flag:"redis-migration-method" cli-usage:"migration method to be used (\"dump\" or \"replication\")" cli-hidden:""`
 	RedisMigrationIgnoreDbs []string `cli-flag:"redis-migration-ignore-dbs" cli-usage:"list of databases which should be ignored during migration" cli-hidden:""`
+
+	// "valkey" type specific flags
+	ValkeyIPFilter           []string `cli-flag:"valkey-ip-filter" cli-usage:"allow incoming connections from CIDR address block" cli-hidden:""`
+	ValkeySettings           string   `cli-flag:"valkey-settings" cli-usage:"Valkey configuration settings (JSON format)" cli-hidden:""`
+	ValkeyMigrationHost      string   `cli-flag:"valkey-migration-host" cli-usage:"hostname or IP address of the source server where to migrate data from" cli-hidden:""`
+	ValkeyMigrationPort      int64    `cli-flag:"valkey-migration-port" cli-usage:"port number of the source server where to migrate data from" cli-hidden:""`
+	ValkeyMigrationPassword  string   `cli-flag:"valkey-migration-password" cli-usage:"password for authenticating to the source server" cli-hidden:""`
+	ValkeyMigrationSSL       bool     `cli-flag:"valkey-migration-ssl" cli-usage:"connect to the source server using SSL" cli-hidden:""`
+	ValkeyMigrationUsername  string   `cli-flag:"valkey-migration-username" cli-usage:"username for authenticating to the source server" cli-hidden:""`
+	ValkeyMigrationDBName    string   `cli-flag:"valkey-migration-dbname" cli-usage:"database name for bootstrapping the initial connection" cli-hidden:""`
+	ValkeyMigrationMethod    string   `cli-flag:"valkey-migration-method" cli-usage:"migration method to be used (\"dump\" or \"replication\")" cli-hidden:""`
+	ValkeyMigrationIgnoreDbs []string `cli-flag:"valkey-migration-ignore-dbs" cli-usage:"list of databases which should be ignored during migration" cli-hidden:""`
 }
 
 func (c *dbaasServiceUpdateCmd) cmdAliases() []string { return nil }
@@ -132,6 +145,9 @@ func (c *dbaasServiceUpdateCmd) cmdPreRun(cmd *cobra.Command, args []string) err
 		os.Exit(0)
 	case cmd.Flags().Changed("help-pg"):
 		cmdShowHelpFlags(cmd.Flags(), "pg-")
+		os.Exit(0)
+	case cmd.Flags().Changed("help-valkey"):
+		cmdShowHelpFlags(cmd.Flags(), "valkey-")
 		os.Exit(0)
 	case cmd.Flags().Changed("help-redis"):
 		cmdShowHelpFlags(cmd.Flags(), "redis-")
@@ -171,6 +187,8 @@ func (c *dbaasServiceUpdateCmd) cmdRun(cmd *cobra.Command, args []string) error 
 		return c.updateMysql(cmd, args)
 	case "pg":
 		return c.updatePG(cmd, args)
+	case "valkey":
+		return c.updateValkey(cmd, args)
 	case "redis":
 		return c.updateRedis(cmd, args)
 	}