+++ cmd/dbaas_user_delete.go
@@ -48,6 +48,8 @@ func (c *dbaasUserDeleteCmd) cmdRun(cmd *cobra.Command, args []string) error {
 		return c.deletePg(cmd, args)
 	case "opensearch":
 		return c.deleteOpensearch(cmd, args)
+	case "valkey":
+		return c.deleteValkey(cmd, args)
 	default:
 		return fmt.Errorf("deleting user unsupported for service of type %q", db.Type)
 	}