+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - storage: bucket replication support #668
 - sks: move cluster creation to egoscale v3 + enable-kube-proxy flag
+- dbaas: added valkey
+- dbaas: remove redis create
 
 ## 1.83.1